First approach to validation has been incorporated.
A Validator instance needs to be provided for this component to work.
The test configuration now provides one in order to run. 

2 spots to run validation.
1. Component uri needs to be validated. Uri parameters are loaded to
CMConfiguration instance. Now this instance is annotated with javax
validation constraints to fulfill CM requirements. 
2. Every immutable SMSMessage received from client needs to be
validated. So now, a validator instance is needed in the CMProducer, and
SMSMessage fields are annotated with javax validation constraints.